# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "efgs-federation-gateway",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0-SNAPSHOT"
  },
  "servers": [{
      "url": "http://localhost:8090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/diagnosiskeys/upload": {
      "post": {
        "tags": ["Diagnosis Keys Exchange Interface", "Upload"],
        "summary": "Uploads diagnosis key datasets.",
        "description": "Uploads the given batch to the server. Uploader Information is given by the client certificate.",
        "operationId": "uploadDiagnosisKeys",
        "parameters": [{
            "name": "batchTag",
            "in": "header",
            "description": "Required Tag to tag the send batch (must be not unique).",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "20200731-1"
          }, {
            "name": "batchSignature",
            "in": "header",
            "description": "PKC7 Payload signature in Base64 encoding.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "ABDBJ345231DJ122..."
          }, {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "application/protobuf; version=1.0"
          }
        ],
        "requestBody": {
          "description": "Requestbody with payload. (limited)",
          "content": {
            "application/protobuf; version=1.0": {},
            "application/json; version=1.0": {}
          }
        },
        "responses": {
          "201": {
            "description": "Database Entries created.",
            "headers": {
              "batchTag": {
                "description": "Tag of the batch.",
                "required": true,
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "207": {
            "description": "Data partially added with warnings. More details in document.",
            "headers": {
              "batchTag": {
                "description": "Tag of the batch.",
                "required": true,
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json+v1.0": {
                "example": "{\n  '201': [1,2,5,8,9],\n  '409': [3,4,6,7],\n  '500': [10]\n}"
              }
            }
          },
          "400": {
            "description": "Signature not valid. Bad request."
          },
          "403": {
            "description": "Forbidden call in cause of missing or invalid client certificate."
          },
          "406": {
            "description": "Data format or content is not valid."
          },
          "413": {
            "description": "Data already exist."
          },
          "500": {
            "description": "Not able to write data. Retry please."
          }
        }
      }
    }
  },
...
}